Use dir() instead of vars() in `Referable.update_from()` (#338)
Previously, we used `vars()` to iterate over all attributes of an object
when updating it from another object (`Referable.update_from()`).

However, `vars()` only retrieves the instance's `__dict__`, e.g. the 
object's attributes, but does not include properties. As a result, when 
you use `vars(other)`, you get the `_id_short` attribute instead of the 
`id_short` property.

This made it impossible to utilize custom setter methods we developed
for certain cases, e.g. when properties should be immutable.

This now adapts the `Referable.update_from()` method to utilize `dir()`
instead, which iterates over all attributes and properties of the 
object. We skip callables and private attributes. We use `setattr()` to 
set attributes and properties, so that the correct setter methods will 
be used.

Furthermore, we now raise an error if an immutable property changed 
between the two versions of the object.

Fixes #215 

---------

Co-authored-by: Sercan Sahin <s.sahin@iat.rwth-aachen.de>