aasx.py: Add failsafe for writing and reading AASX files (#378)
Previously, the `failsafe` parameter, allowing to skip errors and only 
log them as warnings instead, only existed for JSON and XML 
(de-)serialization. 

This PR extends the feature to also be included when reading and 
writing AASX files. For this, we add the `failsafe` boolean attribute 
to the `AASXReader` and `AASXWriter` classes in `adapter.aasx`. 
If `failsafe` is `True`, the document is parsed in a failsafe way: 
Missing attributes and elements are logged instead of causing 
exceptions. Defect objects are skipped. The default value is `False`, 
keeping the existing behavior.

Furthermore, we do a little code-style clean up on string-formatting 
in the `adapter.aasx` module, updating all occurrences from 
(historically) different syntaxes to the more modern f-string syntax.

Fixes #228