sdk\basyx\aas\adapter\xml\xml_serialization.py: Remove code duplication (#406)
Previously, the method `object_to_xml_element()` in
`sdk\basyx\aas\adapter\xml\xml_serialization.py` contained duplicated
code from `submodel_element_to_xml()` and `data_element_to_xml()`. This
was due to the class hierarchy of the objects being compared in the if
statements.

For instance, some models inherit from `model.SubmodelElement`. Because
of that, the associated objects are checked for in
`submodel_element_to_xml()`. With that, any additional check for these
objects in `object_to_xml_element()` are redundant.

This removes the code duplication.

Fixes #396