Refactor `Backend` concept for data persistence  (#370)
Previously, we tried to use the `Backend` concept for both data persistence and data integration. 
This led to complicated cross-dependencies, architectures and badly maintainable code. 

Now, we decided to split these two concepts and use the existing `Backend` `ObjectStore`s for just the data persistence.

Concretely, this commit does the following:

1. Remove the abstract `Backend` class (backends.py), as well as their concrete implementations (`CouchDBBackend`, `LocalFileBackend`), as they are no longer needed
2. Remove the `Referable.source` attribute, as it was only needed for the `Backend` class
3. Remove methods related to the `Referable.source` attribute, such as `Referable.update()` and `Referable.commit()` (base.py), as they were used together to perform the data integration.
4. Adapt other related codes such as tutorials and tests to reflect these changes
5. Adapt documentation to reflect these changes