adapter.aasx: Add support for loading/saving thumbnails in AASX package (#436)
adapter.aasx: Add support for loading and saving thumbnails in AASX package

Previously, the `AASXReader` did not load the thumbnail in the `read_into` function, instead a separate function needed to be called. This behavior did also occur in `AASXWriter` with writing the thumbnail. This was unintuitive, as it required a separat call of the load/save function.

This PR implements the loading/storing of the thumbnail directly in the loading/storing of other supplementary files.

Fixes #435